+++ SmartTray/SmartTray/Models/Requests/TraySensorReadingRequest.cs
@@ -3,9 +3,9 @@
     public class TraySensorReadingRequest
     {
         public DateTime Date { get; set; }
-        public double Temperature { get; set; }
-        public double Humidity { get; set; }
-        public double UvReading { get; set; }
+        public int Temperature { get; set; }
+        public int Humidity { get; set; }
+        public int UvReading { get; set; }
         public bool WaterAdded { get; set; }
         public bool UvLedsOn { get; set; }
     }
