+++ SmartTray/SmartTray/Models/Responses/GrowthSettingsResponse.cs
@@ -3,8 +3,8 @@
     public class GrowthSettingsResponse
     {
         public DateTime RegisterDate { get; set; }
-        public double TemperatureCelsius { get; set; }
-        public double Humidity { get; set; }
-        public double DailySolarHours { get; set; }
+        public int TemperatureCelsius { get; set; }
+        public int Humidity { get; set; }
+        public int DailySolarHours { get; set; }
     }
 }
